GPs and primary care networks
GP Practices and LSOAs - assigning GPs to LSOAs based on highest patient registrations July 2019

Here's the problem.

Coventry and Rugby CCG (CRCCG) is unable to provide boundaries for each Primary Care Network (PCN), because there isn't a simple way to assign GP practices or PCN to a statistical geography.

However, CRCCG wants to be able to divide LSOA data by PCN geography.

The workaround is to assign each GP practice to the best match LSOA.

The method used is as follows:

1. Using the gp-reg-prac-lsoa-all sheet from NHS digital (patients registered at a GP practice July 2019) taken from https://digital.nhs.uk/data-and-information/publications/statistical/patients-registered-at-a-gp-practice/july-2019, for each LSOA, identify the GP practice with the highest patient number. This is a winner-takes-all method where each LSOA gets assigned to one LSOA. Where LSOAs have the same number of patients, then it is at random which GP practice it'll be aasisgned to.

2. Get a list of every single LSOA in the country.

The instructions are as follows:

1. On the LSOA sheet (2.), use the MAXIFS function to identify LSOAs =MAXIFS([patient_numbers],[all_LSOAs],[LSOA_in_question])
2. Concatenate the LSOA code and patient numbers on both the LSOA sheet (2.) and gp-reg-prac-lsoa-all sheet (1.) to get a unique reference number.
3. Use the OFFSET MATCH functions to identify the relevant GP practice with the highest number of patients.
4. Filter out duplicates.